Initial commit
Add configuration files for guilty processing. Files are intended to be
simple to update and push to change guilty processing results. The
version of the file indicates its format, files that change structure
will get a new version and the version processing logic will need be to
updated and then look for the new file name.

Signed-off-by: William Douglas <william.douglas@intel.com>